Version 5.5.0 (#464)
* Adding decoder option for NVEncC and QSVEncC
* Adding option to disable messages on encoding completion or error
* Adding output depth to VCEEncC
* Adding #266 multiple disposition types for subtitles (thanks to lobofoots)
* Adding #326 support for hwaccel decoding with ffmpeg nvenc (thanks to Jim Leonard)
* Adding #329 Hardware VAAPI encoding support for HEVC, H264, VP9 and MPEG2
* Adding #351 option to disable automatic tab switching (thanks to Wontell)
* Adding #357 QSVencC adapt-ref, adapt-ltr, and adapt-qm parameters (thanks to F.O.R.A.R.T.)
* Adding #368 VCEEncC Pre-Analysis options of sc, ss, activity type, caq, sc qp,lookahead, ltr, paq, taq and motion quality (thanks to Wontell)
* Adding #379 #410 dispositions for audio tracks (thanks to markmarz and Don Gafford)
* Fixing videotoolbox setting panels not having custom_q
* Fixing end of run command ran after message box was closed instead of before (thanks to sn6op)
* Fixing changing video or track title did not update the command

+++ fastflix/ff_queue.py
@@ -108,11 +108,15 @@ def save_queue(queue: list[Video], queue_file: Path, config: Optional[Config] = 
                 video["video_settings"]["video_encoder_settings"]["hdr10plus_metadata"] = str(new_metadata_file)
             for track in video["video_settings"]["attachment_tracks"]:
                 if track.get("file_path"):
+                    if not Path(track["file_path"]).exists():
+                        logger.exception("Could not save cover to queue recovery location, removing cover")
+                        continue
                     new_file = queue_covers / f'{uuid.uuid4().hex}_{track["file_path"].name}'
                     try:
                         shutil.copy(track["file_path"], new_file)
                     except OSError:
                         logger.exception("Could not save cover to queue recovery location, removing cover")
+                        continue
                     update_conversion_command(video, str(track["file_path"]), str(new_file))
                     track["file_path"] = str(new_file)
